executor: ajust some code.
