+++ executor/aggregate.go
