+++ executor/aggregate_test.go
@@ -37,10 +37,6 @@ func (m *MockExec) Schema() *expression.Schema {
 	return expression.NewSchema()
 }
 
-func (m *MockExec) Fields() []*ast.ResultField {
-	return m.fields
-}
-
 func (m *MockExec) Next() (*executor.Row, error) {
 	if m.curRowIdx >= len(m.Rows) {
 		return nil, nil
@@ -363,4 +359,43 @@ func (s *testSuite) TestStreamAgg(c *C) {
 			c.Assert(fmt.Sprintf("%v", row.Data[0].GetValue()), Equals, res)
 		}
 	}
+}
+
+func (s *testSuite) TestAggPrune(c *C) {
+	defer func() {
+		s.cleanEnv(c)
+		testleak.AfterTest(c)()
+	}()
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	tk.MustExec("drop table if exists t")
+	tk.MustExec("create table t(id int primary key, b varchar(50), c int)")
+	tk.MustExec("insert into t values(1, '1ff', NULL), (2, '234.02', 1)")
+	tk.MustQuery("select id, sum(b) from t group by id").Check(testkit.Rows("1 1", "2 234.02"))
+	tk.MustQuery("select id, count(c) from t group by id").Check(testkit.Rows("1 0", "2 1"))
+	tk.MustExec("drop table if exists t")
+	tk.MustExec("create table t(id int primary key, b float, c float)")
+	tk.MustExec("insert into t values(1, 1, 3), (2, 1, 6)")
+	tk.MustQuery("select sum(b/c) from t group by id").Check(testkit.Rows("0.3333333333333333", "0.16666666666666666"))
+	tk.MustExec("drop table if exists t")
+	tk.MustExec("create table t(id int primary key, b float, c float, d float)")
+	tk.MustExec("insert into t values(1, 1, 3, NULL), (2, 1, NULL, 6), (3, NULL, 1, 2), (4, NULL, NULL, 1), (5, NULL, 2, NULL), (6, 3, NULL, NULL), (7, NULL, NULL, NULL), (8, 1, 2 ,3)")
+	tk.MustQuery("select count(distinct b, c, d) from t group by id").Check(testkit.Rows("0", "0", "0", "0", "0", "0", "0", "1"))
+}
+
+func (s *testSuite) TestSelectDistinct(c *C) {
+	defer func() {
+		s.cleanEnv(c)
+		testleak.AfterTest(c)()
+	}()
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	s.fillData(tk, "select_distinct_test")
+
+	tk.MustExec("begin")
+	r := tk.MustQuery("select distinct name from select_distinct_test;")
+	rowStr := fmt.Sprintf("%v", []byte("hello"))
+	r.Check(testkit.Rows(rowStr))
+	tk.MustExec("commit")
+
 }
