+++ executor/write.go
